backend/travel/requirements.txt
Django>=4.2,<5.1
wagtail>=6.0,<6.1
yamjam==0.1.7
psycopg2==2.9.9
psycopg2-binary==2.9.9
boto3==1.34.74
django-storages==1.14.2
djangorestframework==3.15.1
wagtail-modeladmin==2.0.0
celery==5.3.6
requests
django-cors-headers==4.3.1